Step0_Api_Basics/profiles-rest-api/requirements.txt
asgiref==3.5.1
Django==3.2.4
numpy==1.22.3
pandas==1.4.2
Pillow==9.1.1
plotly==5.6.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tenacity==8.0.1
whitenoise==6.1.0
djangorestframework